+++ components/connections/protocols/dmx/src/sacn.rs
@@ -24,11 +24,11 @@ impl DmxOutput for SacnOutput {
         format!("sACN ({})", self.source.name())
     }
 
-    fn write_single(&self, universe: u16, channel: u8, value: u8) {
+    fn write_single(&self, universe: u16, channel: u16, value: u8) {
         self.buffer.write_single(universe, channel, value)
     }
 
-    fn write_bulk(&self, universe: u16, channel: u8, values: &[u8]) {
+    fn write_bulk(&self, universe: u16, channel: u16, values: &[u8]) {
         self.buffer.write_bulk(universe, channel, values)
     }
